unit test product order table

+++ src/main/java/com/softlines/fastpos/controller/TablesController.java
@@ -3,18 +3,20 @@ package com.softlines.fastpos.controller;
 import com.softlines.fastpos.domain.Tables;
 import com.softlines.fastpos.dto.TableDto;
 import com.softlines.fastpos.dto.mapping.TableMapper;
+import com.softlines.fastpos.exceptionmanagement.ExceptionManagement;
 import com.softlines.fastpos.repository.TableRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.server.ResponseStatusException;
+
 import java.util.List;
 import java.util.Optional;
 
 @RestController
 @RequestMapping("/table")
 public class TablesController {
+    ExceptionManagement exceptionManagement = new ExceptionManagement();
 
     @Autowired
     TableMapper tableMapper;
@@ -23,33 +25,42 @@ public class TablesController {
     private TableRepository tableRepository;
 
     @PostMapping("/save")
-    public ResponseEntity addTable(@RequestBody Tables table) {
+    public ResponseEntity<TableDto> addTable(@RequestBody TableDto tableDto) {
+
         try {
 
-            Optional<Tables> optionalTable = tableRepository.findById(table.getId());
+            Optional<Tables> optionalTable = tableRepository.findById(tableDto.getId());
 
-            if (!optionalTable.isPresent()) {
-                return ResponseEntity.status(HttpStatus.CREATED).body(tableRepository.save(table));
+            if (!optionalTable.isPresent() && tableDto.getId() == 0) {
+
+                if (tableDto.getNumber() !=0) {
+                    Tables tables = tableMapper.toTable(tableDto);
+                    return ResponseEntity.status(HttpStatus.CREATED).body(tableMapper.toTableDto(tableRepository.save(tables)));
+                }else{
+                    return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+                }
+
             } else {
                 return ResponseEntity.status(HttpStatus.FOUND).build();
-
             }
+
         } catch (Exception exception) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, " Not Found", exception);
+            return exceptionManagement.getResponseEntityAccordingToException(exception);
         }
+
     }
 
     @GetMapping("/getall")
     public ResponseEntity<List<TableDto>> getTables() {
         try {
             List<Tables> tables = tableRepository.findAll();
-            if (tables != null)
-
-                return ResponseEntity.ok().body(tableMapper.toTableDTOs(tables) );
+            if (tables == null || tables.isEmpty())
+                return ResponseEntity.noContent().build();
             else
-                return ResponseEntity.notFound().build();
+                return ResponseEntity.ok().body(tableMapper.toTableDTOs(tables));
+
         } catch (Exception exception) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, " Not Found", exception);
+            return exceptionManagement.getResponseEntityAccordingToException(exception);
         }
 
     }
@@ -60,31 +71,32 @@ public class TablesController {
         try {
             Optional<Tables> optionalTable = tableRepository.findById(id);
 
-            if (optionalTable.isPresent())
+            if (optionalTable.isPresent() && id != 0)
                 return ResponseEntity.ok().body(tableMapper.toTableDto(optionalTable.get()));
             else
-                return ResponseEntity.notFound().build();
+                return ResponseEntity.noContent().build();
 
         } catch (Exception exception) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, " Not Found", exception);
+            return exceptionManagement.getResponseEntityAccordingToException(exception);
         }
     }
 
     @PutMapping("/put/{id}")
-    public ResponseEntity editTable(@PathVariable long id, @RequestBody Tables Table) {
+    public ResponseEntity editTable(@PathVariable long id, @RequestBody TableDto tableDto) {
 
         try {
             Optional<Tables> optionalTable = tableRepository.findById(id);
 
-            if (optionalTable.isPresent()) {
+            if (optionalTable.isPresent() && tableDto.getNumber()!=0) {
 
-                return ResponseEntity.ok().body(tableRepository.save(optionalTable.get()));
 
+                    Tables tables = tableMapper.toTable(tableDto);
+                    return ResponseEntity.ok().body(tableRepository.save(tables));
             } else {
-                return ResponseEntity.notFound().build();
+                return ResponseEntity.noContent().build();
             }
         } catch (Exception exception) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, " Not Found", exception);
+            return exceptionManagement.getResponseEntityAccordingToException(exception);
         }
     }
 
@@ -94,17 +106,20 @@ public class TablesController {
         try {
 
             Optional<Tables> TableToDel = tableRepository.findById(id);
-            if (TableToDel.isPresent()) {
+
+            if (TableToDel.isPresent() && id != 0) {
 
                 tableRepository.delete(TableToDel.get());
                 return ResponseEntity.ok().build();
 
             } else {
+
                 return ResponseEntity.notFound().build();
+
             }
 
         } catch (Exception exception) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, " Not Found", exception);
+            return exceptionManagement.getResponseEntityAccordingToException(exception);
         }
 
     }
